+++ pyres/exceptions.py
@@ -1,2 +1,11 @@
 class NoQueueError(Exception):
     pass
+
+class JobError(RuntimeError):
+    pass
+
+class TimeoutError(JobError):
+    pass
+
+class CrashError(JobError):
+    pass