+++ pyres/job.py
@@ -1,8 +1,10 @@
+import logging
+import time
 from datetime import timedelta
 from pyres import ResQ, safe_str_to_class
 from pyres import failure
 from pyres.failure.redis import RedisBackend
-
+from pyres.compat import string_types
 
 class Job(object):
     """Every job on the ResQ is an instance of the *Job* class.
@@ -24,12 +26,14 @@ class Job(object):
     """
 
     safe_str_to_class = staticmethod(safe_str_to_class)
-    
+
     def __init__(self, queue, payload, resq, worker=None):
         self._queue = queue
         self._payload = payload
         self.resq = resq
         self._worker = worker
+
+        self.enqueue_timestamp = self._payload.get("enqueue_timestamp")
 
         # Set the default back end, jobs can override when we import them
         # inside perform().
@@ -43,7 +47,20 @@ class Job(object):
         """This method converts payload into args and calls the ``perform``
         method on the payload class.
 
-        #@ add entry_point loading
+        Before calling ``perform``, a ``before_perform`` class method
+        is called, if it exists.  It takes a dictionary as an argument;
+        currently the only things stored on the dictionary are the
+        args passed into ``perform`` and a timestamp of when the job
+        was enqueued.
+
+        Similarly, an ``after_perform`` class method is called after
+        ``perform`` is finished.  The metadata dictionary contains the
+        same data, plus a timestamp of when the job was performed, a
+        ``failed`` boolean value, and if it did fail, a ``retried``
+        boolean value.  This method is called after retry, and is
+        called regardless of whether an exception is ultimately thrown
+        by the perform method.
+
 
         """
         payload_class_str = self._payload["class"]
@@ -51,11 +68,35 @@ class Job(object):
         payload_class.resq = self.resq
         args = self._payload.get("args")
 
+        metadata = dict(args=args)
+        if self.enqueue_timestamp:
+            metadata["enqueue_timestamp"] = self.enqueue_timestamp
+
+        before_perform = getattr(payload_class, "before_perform", None)
+
+        metadata["failed"] = False
+        metadata["perform_timestamp"] = time.time()
+        check_after = True
         try:
+            if before_perform:
+                payload_class.before_perform(metadata)
             return payload_class.perform(*args)
-        except:
+        except Exception as e:
+            metadata["failed"] = True
+            metadata["exception"] = e
             if not self.retry(payload_class, args):
+                metadata["retried"] = False
                 raise
+            else:
+                metadata["retried"] = True
+                logging.exception("Retry scheduled after error in %s", self._payload)
+        finally:
+            after_perform = getattr(payload_class, "after_perform", None)
+
+            if after_perform:
+                payload_class.after_perform(metadata)
+
+            delattr(payload_class,'resq')
 
     def fail(self, exception):
         """This method provides a way to fail a job and will use whatever
@@ -68,6 +109,11 @@ class Job(object):
         return fail
 
     def retry(self, payload_class, args):
+        """This method provides a way to retry a job after a failure.
+        If the jobclass defined by the payload containes a ``retry_every`` attribute then pyres
+        will attempt to retry the job until successful or until timeout defined by ``retry_timeout`` on the payload class.
+
+        """
         retry_every = getattr(payload_class, 'retry_every', None)
         retry_timeout = getattr(payload_class, 'retry_timeout', 0)
 
@@ -83,11 +129,13 @@ class Job(object):
         return False
 
     @classmethod
-    def reserve(cls, queue, res, worker=None, timeout=10):
-        """Reserve a job on the queue. This marks this job so that other workers
-        will not pick it up.
+    def reserve(cls, queues, res, worker=None, timeout=10):
+        """Reserve a job on one of the queues. This marks this job so
+        that other workers will not pick it up.
 
         """
-        payload = res.pop(queue, timeout=timeout)
+        if isinstance(queues, string_types):
+            queues = [queues]
+        queue, payload = res.pop(queues, timeout=timeout)
         if payload:
             return cls(queue, payload, res, worker)