+++ tests/__init__.py
@@ -14,7 +14,7 @@ class BasicMulti(object):
     queue = 'basic'
     @staticmethod
     def perform(name, age):
-        print 'name: %s, age: %s' % (name, age)
+        print('name: %s, age: %s' % (name, age))
 
 
 class ReturnAllArgsJob(object):
@@ -37,6 +37,41 @@ class RetryOnExceptionJob(object):
         else:
             return True
 
+class TimeoutJob(object):
+    queue = 'basic'
+
+    @staticmethod
+    def perform(wait_for):
+        import time
+        time.sleep(wait_for)
+        return "Done Sleeping"
+
+class CrashJob(object):
+    queue = 'basic'
+
+    @staticmethod
+    def perform():
+        # Dangerous, this will cause a hard crash of the python process
+        import ctypes
+        ctypes.string_at(1)
+        return "Never got here"
+
+class PrematureExitJob(object):
+    queue = 'basic'
+
+    @staticmethod
+    def perform(exit_code):
+        import sys
+        sys.exit(exit_code)
+        return "Never got here"
+
+class PrematureHardExitJob(object):
+    queue = 'basic'
+
+    @staticmethod
+    def perform(exit_code):
+        os._exit(exit_code)
+        return "Never got here"
 
 class TestProcess(object):
     queue = 'high'
@@ -62,7 +97,7 @@ class LongObject(object):
     def perform(sleep_time):
         import time
         time.sleep(sleep_time)
-        print 'Done Sleeping'
+        print('Done Sleeping')
 
 def test_str_to_class():
     ret = str_to_class('tests.Basic')