+++ tests/test_worker.py
@@ -1,4 +1,4 @@
-from tests import PyResTests, Basic, TestProcess, ErrorObject, RetryOnExceptionJob
+from tests import PyResTests, Basic, TestProcess, ErrorObject, RetryOnExceptionJob, TimeoutJob, CrashJob, PrematureExitJob, PrematureHardExitJob
 from pyres import ResQ
 from pyres.job import Job
 from pyres.scheduler import Scheduler
@@ -13,7 +13,7 @@ class WorkerTests(PyResTests):
         from pyres.exceptions import NoQueueError
         self.assertRaises(NoQueueError, Worker,[])
         self.assertRaises(Exception, Worker,['test'],TestProcess())
-    
+
     def test_startup(self):
         worker = Worker(['basic'])
         worker.startup()
@@ -24,13 +24,13 @@ class WorkerTests(PyResTests):
         assert signal.getsignal(signal.SIGINT) == worker.shutdown_all
         assert signal.getsignal(signal.SIGQUIT) == worker.schedule_shutdown
         assert signal.getsignal(signal.SIGUSR1) == worker.kill_child
-    
+
     def test_register(self):
         worker = Worker(['basic'])
         worker.register_worker()
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         assert self.redis.sismember('resque:workers',name)
-    
+
     def test_unregister(self):
         worker = Worker(['basic'])
         worker.register_worker()
@@ -38,7 +38,7 @@ class WorkerTests(PyResTests):
         assert self.redis.sismember('resque:workers',name)
         worker.unregister_worker()
         assert name not in self.redis.smembers('resque:workers')
-    
+
     def test_working_on(self):
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         self.resq.enqueue(Basic,"test1")
@@ -46,35 +46,35 @@ class WorkerTests(PyResTests):
         worker = Worker(['basic'])
         worker.working_on(job)
         assert self.redis.exists("resque:worker:%s" % name)
-    
+
     def test_processed(self):
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         worker = Worker(['basic'])
         worker.processed()
         assert self.redis.exists("resque:stat:processed")
         assert self.redis.exists("resque:stat:processed:%s" % name)
-        assert self.redis.get("resque:stat:processed") == str(1)
-        assert self.redis.get("resque:stat:processed:%s" % name) == str(1)
+        assert self.redis.get("resque:stat:processed").decode() == str(1)
+        assert self.redis.get("resque:stat:processed:%s" % name).decode() == str(1)
         assert worker.get_processed() == 1
         worker.processed()
-        assert self.redis.get("resque:stat:processed") == str(2)
-        assert self.redis.get("resque:stat:processed:%s" % name) == str(2)
+        assert self.redis.get("resque:stat:processed").decode() == str(2)
+        assert self.redis.get("resque:stat:processed:%s" % name).decode() == str(2)
         assert worker.get_processed() == 2
-    
+
     def test_failed(self):
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         worker = Worker(['basic'])
         worker.failed()
         assert self.redis.exists("resque:stat:failed")
         assert self.redis.exists("resque:stat:failed:%s" % name)
-        assert self.redis.get("resque:stat:failed") == str(1)
-        assert self.redis.get("resque:stat:failed:%s" % name) == str(1)
+        assert self.redis.get("resque:stat:failed").decode() == str(1)
+        assert self.redis.get("resque:stat:failed:%s" % name).decode() == str(1)
         assert worker.get_failed() == 1
         worker.failed()
-        assert self.redis.get("resque:stat:failed") == str(2)
-        assert self.redis.get("resque:stat:failed:%s" % name) == str(2)
+        assert self.redis.get("resque:stat:failed").decode() == str(2)
+        assert self.redis.get("resque:stat:failed:%s" % name).decode() == str(2)
         assert worker.get_failed() == 2
-    
+
     def test_process(self):
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         self.resq.enqueue(Basic,"test1")
@@ -89,8 +89,8 @@ class WorkerTests(PyResTests):
         assert not self.redis.get('resque:worker:%s' % worker)
         assert not self.redis.get("resque:stat:failed")
         assert not self.redis.get("resque:stat:failed:%s" % name)
-        
-    
+
+
     def test_signals(self):
         worker = Worker(['basic'])
         worker.startup()
@@ -104,16 +104,16 @@ class WorkerTests(PyResTests):
         #worker.work()
         #assert worker.child
         assert not worker.kill_child(frame, signal.SIGUSR1)
-    
+
     def test_job_failure(self):
         self.resq.enqueue(ErrorObject)
         worker = Worker(['basic'])
         worker.process()
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         assert not self.redis.get('resque:worker:%s' % worker)
-        assert self.redis.get("resque:stat:failed") == str(1)
-        assert self.redis.get("resque:stat:failed:%s" % name) == str(1)
-    
+        assert self.redis.get("resque:stat:failed").decode() == str(1)
+        assert self.redis.get("resque:stat:failed:%s" % name).decode() == str(1)
+
     def test_get_job(self):
         worker = Worker(['basic'])
         self.resq.enqueue(Basic,"test1")
@@ -122,11 +122,11 @@ class WorkerTests(PyResTests):
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
         assert worker.job() == ResQ.decode(self.redis.get('resque:worker:%s' % name))
         assert worker.processing() == ResQ.decode(self.redis.get('resque:worker:%s' % name))
-        worker.done_working()
+        worker.done_working(job)
         w2 = Worker(['basic'])
-        print w2.job()
+        print(w2.job())
         assert w2.job() == {}
-    
+
     def test_working(self):
         worker = Worker(['basic'])
         self.resq.enqueue_from_string('tests.Basic','basic','test1')
@@ -138,18 +138,18 @@ class WorkerTests(PyResTests):
         assert len(workers) == 1
         assert str(worker) == str(workers[0])
         assert worker != workers[0]
-    
+
     def test_started(self):
         import datetime
         worker = Worker(['basic'])
         dt = datetime.datetime.now()
         worker.started = dt
         name = "%s:%s:%s" % (os.uname()[1],os.getpid(),'basic')
-        assert self.redis.get('resque:worker:%s:started' % name) == str(int(time.mktime(dt.timetuple())))
-        assert worker.started == str(int(time.mktime(dt.timetuple())))
+        assert self.redis.get('resque:worker:%s:started' % name).decode() == str(int(time.mktime(dt.timetuple())))
+        assert worker.started.decode() == str(int(time.mktime(dt.timetuple())))
         worker.started = None
         assert not self.redis.exists('resque:worker:%s:started' % name)
-    
+
     def test_state(self):
         worker = Worker(['basic'])
         assert worker.state() == 'idle'
@@ -158,9 +158,9 @@ class WorkerTests(PyResTests):
         job = Job.reserve('basic', self.resq)
         worker.working_on(job)
         assert worker.state() == 'working'
-        worker.done_working()
+        worker.done_working(job)
         assert worker.state() == 'idle'
-    
+
     def test_prune_dead_workers(self):
         worker = Worker(['basic']) # we haven't registered this worker, so the assertion below holds
         assert self.redis.scard('resque:workers') == 0
@@ -202,6 +202,76 @@ class WorkerTests(PyResTests):
         assert True == worker.process()
         assert worker.get_failed() == 0
 
+    def test_kills_stale_workers_after_timeout(self):
+        timeout = 1
+
+        worker = Worker(['basic'], timeout=timeout)
+        self.resq.enqueue(TimeoutJob, timeout + 1)
+
+        assert worker.get_failed() == 0
+        worker.fork_worker(worker.reserve())
+        assert worker.get_failed() == 1
+
+    def test_detect_crashed_workers_as_failures(self):
+        worker = Worker(['basic'])
+        self.resq.enqueue(CrashJob)
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
+        worker.fork_worker(worker.reserve())
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 1
+
+    def test_detect_non_0_sys_exit_as_failure(self):
+        worker = Worker(['basic'])
+        self.resq.enqueue(PrematureExitJob, 9)
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
+        worker.fork_worker(worker.reserve())
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 1
+
+    def test_detect_code_0_sys_exit_as_success(self):
+        worker = Worker(['basic'])
+        self.resq.enqueue(PrematureExitJob, 0)
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
+        worker.fork_worker(worker.reserve())
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
+    def test_detect_non_0_os_exit_as_failure(self):
+        worker = Worker(['basic'])
+        self.resq.enqueue(PrematureHardExitJob, 9)
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
+        worker.fork_worker(worker.reserve())
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 1
+
+    def test_detect_code_0_os_exit_as_success(self):
+        worker = Worker(['basic'])
+        self.resq.enqueue(PrematureHardExitJob, 0)
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
+        worker.fork_worker(worker.reserve())
+
+        assert worker.job() == {}
+        assert worker.get_failed() == 0
+
     def test_retries_give_up_eventually(self):
         now = datetime.datetime.now()
         self.set_current_time(now)
@@ -225,6 +295,37 @@ class WorkerTests(PyResTests):
         scheduler.handle_delayed_items()
         assert None == worker.process()
         assert worker.get_failed() == 1
+
+    def test_worker_pids(self):
+        # spawn worker processes and get pids
+        pids = []
+        pids.append(self.spawn_worker(['basic']))
+        pids.append(self.spawn_worker(['basic']))
+        time.sleep(1)
+        worker_pids = Worker.worker_pids()
+
+        # send kill signal to workers and wait for them to exit
+        import signal
+        for pid in pids:
+            os.kill(pid, signal.SIGQUIT)
+            os.waitpid(pid, 0)
+
+        # ensure worker_pids() returned the correct pids
+        for pid in pids:
+            assert str(pid) in worker_pids
+
+        # ensure the workers are no longer returned by worker_pids()
+        worker_pids = Worker.worker_pids()
+        for pid in pids:
+            assert str(pid) not in worker_pids
+
+    def spawn_worker(self, queues):
+        pid = os.fork()
+        if not pid:
+            Worker.run(queues, interval=1)
+            os._exit(0)
+        else:
+            return pid
 
     def set_current_time(self, time):
         ResQ._current_time = staticmethod(lambda: time)